auralink-ai/backend/requirements.txt
# AuraLink AI - FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Vision / AI
google-generativeai>=0.8.0
openai>=1.50.0
google-cloud-vision>=3.7.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Database (Supabase / PostgreSQL)
supabase>=2.10.0
asyncpg>=0.30.0

# Worker queue
celery[redis]>=5.4.0
redis>=5.2.0

# Auth (verify Clerk JWT)
pyjwt[crypto]>=2.9.0

# HTTP client (Shopify OAuth, GraphQL)
httpx>=0.27.0

# Utils
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
